requirements.txt
python-docx
language-tool-python
requests
python-dotenv
Pillow 